[feat-#30] 회원가입 로직을 auth 패키지로 이동
