+++ src/main/java/com/bss/bssserverapi/domain/user/controller/UserController.java
@@ -1,14 +1,14 @@
 package com.bss.bssserverapi.domain.user.controller;
 
 import com.bss.bssserverapi.domain.user.dto.GetUserResDto;
-import com.bss.bssserverapi.domain.user.dto.SignupUserReqDto;
 import com.bss.bssserverapi.domain.user.service.UserService;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/v1/users")
@@ -16,14 +16,6 @@ import org.springframework.web.bind.annotation.*;
 public class UserController {
 
     private final UserService userService;
-
-    @PostMapping("/signup")
-    public ResponseEntity<?> signupUser(@RequestBody @Valid final SignupUserReqDto signupUserReqDto){
-
-        return ResponseEntity
-                .status(HttpStatus.CREATED)
-                .body(userService.signupUser(signupUserReqDto));
-    }
 
     @GetMapping("/me")
     public ResponseEntity<GetUserResDto> getUser(@AuthenticationPrincipal final String userName) {
