+++ src/test/java/com/bss/bssserverapi/user/UserServiceTest.java
@@ -1,18 +1,10 @@
 package com.bss.bssserverapi.user;
 
-import com.bss.bssserverapi.domain.user.User;
-import com.bss.bssserverapi.domain.user.dto.SignupUserReqDto;
-import com.bss.bssserverapi.domain.user.dto.SignupUserResDto;
 import com.bss.bssserverapi.domain.user.repository.UserJpaRepository;
 import com.bss.bssserverapi.domain.user.service.UserService;
-import com.bss.bssserverapi.global.exception.ErrorCode;
-import com.bss.bssserverapi.global.exception.GlobalException;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -34,46 +26,46 @@ public class UserServiceTest {
     @InjectMocks
     private UserService userService;
 
-    @Test
-    @DisplayName("회원 가입 성공")
-    public void createUser_Success(){
+//    @Test
+//    @DisplayName("회원 가입 성공")
+//    public void createUser_Success(){
+//
+//        // given
+//        SignupUserReqDto req = SignupUserReqDto.builder()
+//                .userName("bss_admin")
+//                .password("Qq12341234@")
+//                .passwordConfirmation("Qq12341234@")
+//                .build();
+//
+//        Mockito.doReturn(User.builder()
+//                .userName("bss_admin")
+//                .password(bCryptPasswordEncoder.encode("Qq12341234@"))
+//                .build())
+//                .when(userJpaRepository)
+//                .save(any(User.class));
+//
+//        // when
+//        SignupUserResDto res = userService.signupUser(req);
+//
+//        // then
+//        assertThat(res.getUserName()).isEqualTo(req.getUserName());
+//    }
 
-        // given
-        SignupUserReqDto req = SignupUserReqDto.builder()
-                .userName("bss_admin")
-                .password("Qq12341234@")
-                .passwordConfirmation("Qq12341234@")
-                .build();
-
-        Mockito.doReturn(User.builder()
-                .userName("bss_admin")
-                .password(bCryptPasswordEncoder.encode("Qq12341234@"))
-                .build())
-                .when(userJpaRepository)
-                .save(any(User.class));
-
-        // when
-        SignupUserResDto res = userService.signupUser(req);
-
-        // then
-        assertThat(res.getUserName()).isEqualTo(req.getUserName());
-    }
-
-    @Test
-    @DisplayName("회원 가입 실패 - 비밀번호, 비밀번호 확인 일치 하지 않음")
-    public void createUser_Fail_PasswordMismatch(){
-
-        // given
-        SignupUserReqDto req = SignupUserReqDto.builder()
-                .userName("bss_admin")
-                .userName("bss_admin")
-                .password("Qq12341234@!!!!!!!!!!!!")
-                .passwordConfirmation("Qq12341234@")
-                .build();
-
-        // when & then
-        assertThatThrownBy(() -> userService.signupUser(req))
-                .isInstanceOf(GlobalException.class)
-                .hasMessage(ErrorCode.PASSWORD_AND_CONFIRMATION_MISMATCH.getMessage());
-    }
+//    @Test
+//    @DisplayName("회원 가입 실패 - 비밀번호, 비밀번호 확인 일치 하지 않음")
+//    public void createUser_Fail_PasswordMismatch(){
+//
+//        // given
+//        SignupUserReqDto req = SignupUserReqDto.builder()
+//                .userName("bss_admin")
+//                .userName("bss_admin")
+//                .password("Qq12341234@!!!!!!!!!!!!")
+//                .passwordConfirmation("Qq12341234@")
+//                .build();
+//
+//        // when & then
+//        assertThatThrownBy(() -> userService.signupUser(req))
+//                .isInstanceOf(GlobalException.class)
+//                .hasMessage(ErrorCode.PASSWORD_AND_CONFIRMATION_MISMATCH.getMessage());
+//    }
 }
